Issue #2359213: Check access for private files.
A few things happened here. I split out the access-checking logic into a
FillPdfAccessHelper service. In coordination with that, I added
FillPdfLinkManipulator::parseUrlString() to facilitate parsing the
context strings we store in FillPdfFileContext objects. I also changed
the way I store URLs in FillPdfFileContext.context so that they are
easier to pass to Url::fromUri() for parsing (and I made a helper
method that parses them into a form Uri::fromUri() likes).

In short, saving PDFs to files now works as well as downloading. Score!